ai-service/requirements.txt
flask
ultralytics
pillow
opencv-python-headless
